requirements.txt
anthropic>=0.40.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
googlesearch-python>=1.2.0
